RELEASE CANDIDATE 2

+++ GameLibrary/src/Time.cs
@@ -4,13 +4,13 @@ namespace GameLibrary;
 
 public static class Time
 {
-    public static double Scaled; // Seconds of game time since last reset. Reset happens at start of battle
-    public static double Unscaled; // Seconds since last reset. Is not affected by pause or ffw, but is affected by lag.
+    public static float Scaled; // Seconds of game time since last reset. Reset happens at start of battle
+    public static float Unscaled; // Seconds since last reset. Is not affected by pause or ffw, but is affected by lag.
     public static int Tick; // Number of updates since last reset.
     public static int Frame; // Number of frames since last reset. Equivalent to Unscaled.
-    public static double Real; // Seconds since program started
+    public static float Real; // Seconds since program started
     public const int FrameRate = 60;
-    public const double DeltaTime = 1.0/FrameRate;
+    public const float DeltaTime = 1.0f/FrameRate;
     public static bool Paused;
     
     public static void UpdateTime(bool fastForward = false)
@@ -25,7 +25,7 @@ public static class Time
             Tick++;
             Scaled = Tick * DeltaTime;
         }
-        Real = Raylib.GetTime();
+        Real = (float)Raylib.GetTime();
     }
 
     public static void Reset()
